backend/app/models/mudra_model.py
import numpy as np
from PIL import Image

class MudraClassifier:
    def __init__(self):
        # Placeholder for model loading
        # self.model = load_your_model_here()
        self.mudra_database = {
            "Pataka": {
                "meaning": "Flag - Represents the beginning of dance, clouds, forest",
                "description": "All fingers extended and held together"
            },
            "Tripataka": {
                "meaning": "Three parts of flag - Represents crown, tree, fire",
                "description": "Pataka with ring finger bent"
            },
            # Add more mudras here
        }
    
    def analyze_image(self, image: Image.Image) -> dict:
        """
        Analyze image and return mudra classification results
        """
        # Placeholder implementation - replace with actual ML model
        # This is where you'll integrate your TensorFlow/PyTorch model
        
        # For now, return mock data
        return {
            "name": "Pataka",
            "accuracy": 85.5,
            "meaning": self.mudra_database["Pataka"]["meaning"],
            "feedback": "Good form! Keep fingers straight and together."
        }
    
    def analyze_practice(self, target_mudra: str, image: Image.Image) -> dict:
        """
        Analyze practice attempt against target mudra
        """
        # Placeholder implementation
        accuracy = np.random.randint(60, 95)
        
        feedbacks = {
            "high": "Excellent! Perfect mudra form.",
            "medium": "Good attempt. Work on finger alignment.",
            "low": "Needs practice. Focus on hand shape."
        }
        
        if accuracy > 80:
            feedback = feedbacks["high"]
        elif accuracy > 65:
            feedback = feedbacks["medium"]
        else:
            feedback = feedbacks["low"]
        
        return {
            "accuracy": accuracy,
            "feedback": feedback
        }
    
    def preprocess_image(self, image: Image.Image) -> np.ndarray:
        """
        Preprocess image for model input
        """
        # Resize, normalize, etc.
        image = image.resize((224, 224))  # Example size
        image_array = np.array(image) / 255.0
        return np.expand_dims(image_array, axis=0)